Update Activity 2.py

+++ Assignment 3/Activity 2.py
@@ -1,6 +1,6 @@
 # This program used to asking user age.
 print("Enter age:")
-age = int(input())
+age = float(input())
 month = age * 12
 day = age * 12 * 365 + float(age) / 4
 hour = (age * 12 * 365 + float(age) / 4) * 24
